requirements.txt
dash==2.17.0
pandas>=2.2.3
supabase==2.5.0
python-dotenv==1.0.1
plotly==5.22.0
gunicorn==21.2.0